% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFold.R
\name{KFold}
\alias{KFold}
\title{K-Folds cross validation index generator}
\usage{
KFold(target, nfolds = 10, stratified = FALSE, seed = 0)
}
\arguments{
\item{target}{Samples to split in K folds.}

\item{nfolds}{Number of folds.}

\item{stratified}{whether to apply Stratified KFold.}

\item{seed}{random seed to be used.}
}
\value{
a list of indices for K-Folds Cross-Validation
}
\description{
Generates a list of indices for K-Folds Cross-Validation.
}
