% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Matrix_runif}
\alias{Matrix_runif}
\title{Matrix runif}
\usage{
Matrix_runif(n, lower, upper)
}
\arguments{
\item{n}{number of observations}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\value{
a matrix of original hyperparameters
}
\description{
Generates random deviates for multiple hyperparameters in matrix format.
}
\keyword{internal}
