% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Min_Max_Inverse_Scale_Vec}
\alias{Min_Max_Inverse_Scale_Vec}
\title{MinMax Inverse Scaling}
\usage{
Min_Max_Inverse_Scale_Vec(vec, lower, upper)
}
\arguments{
\item{lower}{lower bounds}

\item{upper}{upper bounds}

\item{mat}{a vector of scaled hyperparameters}
}
\value{
a vector of original hyperparameters
}
\description{
Transforms scaled hyperparameters to original range.
}
\keyword{internal}
