% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Min_Max_Scale_Mat}
\alias{Min_Max_Scale_Mat}
\title{Matrix MinMax Scaling}
\usage{
Min_Max_Scale_Mat(mat, lower, upper)
}
\arguments{
\item{mat}{a matrix of original hyperparameters}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\value{
a matrix of scaled hyperparameters
}
\description{
Transforms hyperparameters by scaling each hyperparameter to a given range.
}
\keyword{internal}
