% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBayesianOptimization.R
\docType{package}
\name{rBayesianOptimization}
\alias{rBayesianOptimization}
\alias{rBayesianOptimization-package}
\title{rBayesianOptimization: Bayesian Optimization of Hyperparameters}
\description{
An Pure R implementation of bayesian global optimization with gaussian processes.
}

