% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Utility}
\alias{Utility}
\title{Utility Computing Function}
\usage{
Utility(x_vec, GP, acq = "ucb", y_max, kappa, eps)
}
\arguments{
\item{x_vec}{a vector of scaled hyperparameters}

\item{GP}{an object of class GP}

\item{acq}{Acquisition function type to be used}

\item{y_max}{The current maximum known value of the target utility function}

\item{kappa}{tunable parameter kappa to balance exploitation against exploration}

\item{eps}{tunable parameter epsilon to balance exploitation against exploration}
}
\value{
negative utility to be minimized
}
\description{
Computing Utility.
}
\keyword{internal}
